Fix firefox failing build (#217)
Hi,

After some experimentation and digging in the firefox repository I realized that the commas in _link-arg_ are the cause of the error.

Also, correct me if I'm wrong here, but you do not have to pass -Wl to the linker, cargo handles that. Plus, according to the [documentation](https://doc.rust-lang.org/rustc/codegen-options/index.html#link-arg), it is possible to pass _link-arg_ multiple times to achieve the same result.

I managed to build firefox locally with this change, so I guess this should work here as well.

Reviewed-on: https://somegit.dev/ALHP/ALHP.GO/pulls/217
Co-authored-by: strptrk <git@patrikstier.com>
Co-committed-by: strptrk <git@patrikstier.com>